backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
python-dotenv
pydantic
pydantic-settings
pydantic[email]
alembic
python-multipart
aiofiles
bcrypt==3.2.0
